collegemart/requirements.txt
astroid==2.0.4
Babel==2.6.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.3.0
Django==2.1.3
django-phonenumber-field==2.1.0
enum34==1.1.6
InvoiceGenerator==1.0.1
ipython-genutils==0.2.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pathlib2==2.3.2
pexpect==4.6.0
phonenumbers==8.9.16
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycrypto==2.6.1
Pygments==2.2.0
pylint==2.1.1
pytz==2018.7
qrcode==6.0
qrplatba==0.3.4
reportlab==3.5.10
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typed-ast==1.1.0
wcwidth==0.1.7
wrapt==1.10.11
